smart-traffic/oracle/requirements.txt
# Oracle Voice Assistant — Python dependencies
# Install with: pip install -r requirements.txt

# Core AI
anthropic>=0.28.0          # Claude (primary brain)
openai>=1.14.0             # GPT-4o-mini (fallback brain)

# Speech-to-Text
openai-whisper>=20231117   # local Whisper transcription

# Text-to-Speech
gTTS>=2.5.0                # Google TTS (online)
pyttsx3>=2.90              # offline TTS fallback

# Audio I/O
sounddevice>=0.4.6
scipy>=1.12.0
numpy>=1.26.0

# HTTP
requests>=2.31.0

# Config
python-dotenv>=1.0.0

# Optional: for Flask integration in Daarideepa
# flask>=3.0.0
